Refactored example code.
