+++ examples/go/store/run_server.sh
@@ -1,0 +1,7 @@
+#!/bin/sh
+if [ -z $LIGHT_STEP_ACCESS_TOKEN ]
+then
+  echo "LIGHT_STEP_ACCESS_TOKEN must be set"
+  exit -1
+fi
+go run server/main.go --access_token $LIGHT_STEP_ACCESS_TOKEN "$@"
